+++ app/index.js
@@ -105,13 +105,13 @@ require("./scripts/search_controllers.js")
 
 require("./scripts/kwic_download.js")
 
-require("./scripts/results/controllers/result_controllers.js")
-require("./scripts/results/controllers/comparison_controller.js")
-require("./scripts/results/controllers/kwic_controller.js")
-require("./scripts/results/controllers/example_controller.js")
-require("./scripts/results/controllers/statistics_controller.js")
-require("./scripts/results/controllers/trend_diagram_controller.js")
-require("./scripts/results/controllers/word_picture_controller.js")
+require("./scripts/controllers/result_controllers.js")
+require("./scripts/controllers/comparison_controller.js")
+require("./scripts/controllers/kwic_controller.js")
+require("./scripts/controllers/example_controller.js")
+require("./scripts/controllers/statistics_controller.js")
+require("./scripts/controllers/trend_diagram_controller.js")
+require("./scripts/controllers/word_picture_controller.js")
 
 require("./scripts/map_controllers.js")
 require("./scripts/text_reader_controller.js")
@@ -120,7 +120,6 @@ require("./scripts/services.js")
 require("./scripts/extended.js")
 require("./scripts/struct_services.js")
 require("./scripts/directives.js")
-require("./scripts/directives2.js")
 require("./scripts/directives/scroll.js")
 require("./scripts/filter_directives.js")
 require("./scripts/newsdesk.js")
