+++ app/scripts/controllers/comparison_controller.js
