+++ app/scripts/controllers/example_controller.js
