+++ app/scripts/controllers/trend_diagram_controller.js
