+++ app/scripts/controllers/word_picture_controller.js
